# docs/swagger.yaml
basePath: /
...
info:
  contact: {}
  description: API для Emplacc.
  title: Emplacc API
  version: "1.0"
paths:
...
  /upload/refresh:
    get:
      description: Генерирует новый presigned URL для объекта по его пути
      parameters:
      - description: 'Путь к объекту (например: images/uuid.jpg)'
        in: query
        name: path
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: url
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Обновить presigned URL для файла
      tags:
      - Upload
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"